The ATA.BoxCoxAttr function works with many different types of inputs.

The ATA.BoxCoxAttr function works with many different types of inputs.

Usage

ATA.BoxCoxAttr(
  bcMethod = "guerrero",
  bcLower = 0,
  bcUpper = 5,
  bcBiasAdj = FALSE
)

Arguments

bcMethod

Choose method to be used in calculating lambda. "guerrero" (Guerrero, V.M. (1993) is default. Other method is "loglik").

bcLower

Lower limit for possible lambda values. The lower value is limited by -5. Default value is 0.

bcUpper

Upper limit for possible lambda values. The upper value is limited by 5. Default value is 5.

bcBiasAdj

Use adjusted back-transformed mean for Box-Cox transformations. If transformed data is used to produce forecasts and fitted values, a regular back transformation will result in median forecasts. If bcBiasAdj is TRUE, an adjustment will be made to produce mean forecasts and fitted values. If bcBiasAdj=TRUE. Can either be the forecast variance, or a list containing the interval level, the corresponding upper and lower intervals.

Value

An object of class ataoptim.

References

#'Box GEP, Cox DR (1964). “An Analysis of Transformations.” Journal of the Royal Statistical Society. Series B (Methodological), 26(2), 211–252.

#'Guerrero VM (1993). “Time-series analysis supported by power transformations.” Journal of Forecasting, 12(1), 37–48.

See also

Author

Ali Sabri Taylan and Hanife Taylan Selamlar