The core algorithm of the ATA Method

The core algorithm of the ATA Method

ATA.Core(X, pk, qk, phik, mdlType, initialLevel, initialTrend)

Arguments

X
A numeric vector or time series.

pk
Value of Level parameter.

qk
Value of Trend parameter.

phik
Value of Damping Trend parameter.

mdlType
An one-character string identifying method using the framework terminology.

initialLevel
If NULL, FALSE is default. If FALSE, ATA Method calculates the pth observation in X for level. If TRUE, ATA Method calculates average of first p value in Xfor level.

initialTrend
If NULL, FALSE is default. If FALSE, ATA Method calculates the qth observation in X(T)-X(T-1) for trend. If TRUE, ATA Method calculates average of first q value in X(T)-X(T-1) for trend.

Value

Returns an object of class "ATA"