The core algorithm of the ATA Method

The core algorithm of the ATA Method

ATA.Core(X, pk, qk, phik, mdlType, initialLevel, initialTrend)

Arguments

X

A numeric vector or time series.

pk

Value of Level parameter.

qk

Value of Trend parameter.

phik

Value of Damping Trend parameter.

mdlType

An one-character string identifying method using the framework terminology.

initialLevel

"none" is default,

none : ATA Method calculates the pth observation in X for level.

mean : ATA Method calculates average of first p value in Xfor level.

median: ATA Method calculates median of first p value in Xfor level.

initialTrend

"none" is default,

none : ATA Method calculates the qth observation in X for trend.

mean : ATA Method calculates average of first q value in X(T)-X(T-1) for trend.

median: ATA Method calculates median of first q value in X(T)-X(T-1) for trend.

Value

Returns an object of class "ATA"