Seasonal Decomposition for The ATAforecasting

Automatic seasonal decomposition for ATA Method is called ATA.Decomposition function in ATAforecasting package. The function returns seasonally adjusted data constructed by removing the seasonal component. The methodology is fully automatic. The ATA.Decomposition function works with many different types of inputs.

ATA.Decomposition(input, s.model, s.type, s.frequency, seas_attr_set)

Arguments

input
It must be ts or msts or numeric object. if it is numeric object, findPeriod must be 1 or 2 or 3 or 4. if it is msts object, findPeriod must be 3 or 4.

s.model
A string identifying method for seasonal decomposition. If NULL, "decomp" method is default. c("none", "decomp", "stl", "stlplus", "tbats", "stR") phrases of methods denote.

none : seasonal decomposition is not required.

decomp : classical seasonal decomposition. If decomp, the stats package will be used.

stl : seasonal-trend decomposition procedure based on loess developed by Cleveland et al. (1990). If stl, the stats and forecast packages will be used. Multiple seasonal periods are allowed.

stlplus : seasonal-trend decomposition procedure based on loess developed by Cleveland et al. (1990). If stlplus, the stlplus package will be used.

tbats : exponential smoothing state space model with Box--Cox transformation, ARMA errors, trend and seasonal components. as described in De Livera, Hyndman & Snyder (2011). Parallel processing is used by default to speed up the computations. If tbats, the forecast package will be used. Multiple seasonal periods are allowed.

stR : seasonal-trend decomposition procedure based on regression developed by Dokumentov and Hyndman (2015). If stR, the stR package will be used. Multiple seasonal periods are allowed.

x13 : seasonal-trend decomposition procedure based on X13ARIMA/SEATS. If x13, the seasonal package will be used.

x11 : seasonal-trend decomposition procedure based on X11. If x11, the seasonal package will be used.

s.type
A one-character string identifying method for the seasonal component framework. If NULL, "M" is default. The letter "A" for additive model, the letter "M" for multiplicative model.

s.frequency
Value(s) of seasonal periodicity. If s.frequency is not integer, X must be msts time series object. c(s1,s2,s3,...) for multiple period. If X has multiple periodicity, "tbats" or "stR" seasonal model have to be selected.

seas_attr_set
Assign from ATA.SeasAttr function. Attributes set for unit root and seasonality tests. For example: period of the input data which have one seasonal pattern --> 12 for monthly / 4 for quarterly / 7 for daily / 5 for business days. periods of the input data which have complex/multiple seasonal patterns --> c(7,354.37,365.25).

Value

Seasonal components of the univariate time series. ATA.Decomposition is a list containing at least the following elements:

AdjustedX

Deseasonalized data

SeasIndex

Particular weights of seasonality given cycle/frequency

SeasActual

Seasonality given original data

SeasType

Seasonal decomposition technique

References

#'Shiskin J, Young AH, Musgrave JC (1967). “The X-11 Variant of the Census-II Method Seasonal Adjustment Program.” Technical Report 15, Bureau of the U.S. Census. https://www.census.gov/ts/papers/ShiskinYoungMusgrave1967.pdf.

#'Dagum EB (1999). X11ARIMA/2000 An Updated of The X11ARIMA/88 Seasonal Adjustment Method - Foundations and Users' Manual. Statistics Canada. https://www.census.gov/ts/papers/Emanual.pdf.

#'Cleveland RB, Cleveland WS, McRae JE, Terpenning I (1990). “STL: A seasonal-trend decomposition procedure based on loess.” Journal of Official Statistics, 6(1), 3--73.

#'Hafen RP (2010). Local regression models: Advancements, applications, and new methods. Ph.D. thesis, Purdue University.

#'Livera AMD, Hyndman RJ, Snyder RD (2011). “Forecasting Time Series With Complex Seasonal Patterns Using Exponential Smoothing.” Journal of the American Statistical Association, 106(496), 1513--1527.

#'Dokumentov A, Hyndman RJ (2015). “STR: A Seasonal-Trend Decomposition Procedure Based on Regression.” Monash Econometrics and Business Statistics Working Papers 13/15, Monash University, Department of Econometrics and Business Statistics. https://EconPapers.repec.org/RePEc:msh:ebswps:2015-13.

#'Dokumentov A, Hyndman RJ (2020). “STR: A Seasonal-Trend Decomposition Procedure Based on Regression.” 2009.05894.

#'Monsell BC, Aston JAD, Koopman SJ (2003). “Toward X-13?” United States Census Bureau. https://www.census.gov/content/dam/Census/library/working-papers/2003/adrm/jsm2003bcm.pdf.

#'Monsell BC (2007). “The X-13A-S seasonal adjustment program.” In Proceedings of the 2007 Federal Committee On Statistical Methodology Research Conference. URL http://www. fcsm. gov/07papers/Monsell. II-B. pdf.

#'Sax C, Eddelbuettel D (2018). “Seasonal Adjustment by X-13ARIMA-SEATS in R.” Journal of Statistical Software, 87(11), 1--17.

See also

Author

Ali Sabri Taylan and Hanife Taylan Selamlar