Attributes Set For Unit Root and Seasonality Tests

This function is a class of seasonality tests using corrgram.test from ATAforecasting package, ndiffs and nsdiffs functions from forecast package. Also, this function is modified version of ndiffs and nsdiffs written by Hyndman et al. forecast package. Please review manual and vignette documents of latest forecast package. According to forecast package, ndiffs and nsdiffs functions to estimate the number of differences required to make a given time series stationary. ndiffs uses unit root tests to determine the number of differences required for time series to be made trend stationary. Several different tests are available:

uroot.test = 'kpss' : the KPSS test is used with the null hypothesis that x has a stationary root against a unit-root alternative. Then the test returns the least number of differences required to pass the test at the level uroot.alpha.

uroot.test = 'adf' : the Augmented Dickey-Fuller test is used.

uroot.test = 'pp' : the Phillips-Perron test is used. In both of these cases, the null hypothesis is that x has a unit root against a stationary root alternative. Then the test returns the least number of differences required to fail the test at the level alpha.

nsdiffs uses seasonal unit root tests to determine the number of seasonal differences required for time series to be made stationary. Several different tests are available:

suroot.test = 'seas' : a measure of seasonal strength is used, where differencing is selected if the seasonal strength (Wang, Smith & Hyndman, 2006) exceeds 0.64 (based on minimizing MASE when forecasting using auto.arima on M3 and M4 data).

suroot.test = 'ch' : the Canova-Hansen (1995) test is used (with null hypothesis of deterministic seasonality)

suroot.test = 'hegy' : the Hylleberg, Engle, Granger & Yoo (1990) test is used.

suroot.test = 'ocsb' : the Osborn-Chui-Smith-Birchenhall (1988) test is used (with null hypothesis that a seasonal unit root exists).

suroot.test = 'correlogram' : this function is written based on M4 Competition Seasonality Test.

ATA.SeasAttr(corrgram.tcrit = 1.28, uroot.test = "adf",
  suroot.test = "correlogram", suroot.uroot = TRUE,
  uroot.type = "trend", uroot.alpha = 0.05, suroot.alpha = 0.05,
  uroot.maxd = 2, suroot.maxD = 1, suroot.m = NULL,
  uroot.pkg = "tseries", multi.period = "min",
  x13.estimate.maxiter = 1500, x13.estimate.tol = 1e-05,
  x11.estimate.maxiter = 1500, x11.estimate.tol = 1e-05)

Arguments

corrgram.tcrit
t-value for autocorrelogram.

uroot.test
Type of unit root test before all type seasonality test. Possible values are "adf", "pp" and "kpss".

suroot.test
Type of seasonal unit root test to use. Possible values are "correlogram", "seas", "hegy", "ch" and "ocsb".

suroot.uroot
If TRUE, unit root test for stationary before seasonal unit root test is allowed.

uroot.type
Specification of the deterministic component in the regression for unit root test. Possible values are "level" and "trend".

uroot.alpha
Significant level of the unit root test, possible values range from 0.01 to 0.1

suroot.alpha
Significant level of the seasonal unit root test, possible values range from 0.01 to 0.1

uroot.maxd
Maximum number of non-seasonal differences allowed.

suroot.maxD
Maximum number of seasonal differences allowed.

suroot.m
Deprecated. Length of seasonal period: frequency of data for nsdiff.

uroot.pkg
Using ucra or tseries packages for unit root test. The default value is "ucra".

multi.period
Selection type of multi seasonal period. min or max function for selection

x13.estimate.maxiter
Maximum iteration for X13ARIMA/SEATS estimation

x13.estimate.tol
Convergence tolerence for X13ARIMA/SEATS estimation

x11.estimate.maxiter
Maximum iteration for X11 estimation

x11.estimate.tol
Convergence tolerence for X11 estimation

Value

An object of class ataattrset.

See also