Transformation Techniques for ATA Method

These functions are modified version of trafo::trafos written by Lily Medina, Piedad Castro, Ann-Kristin Kreutzmann, Natalia Rojas-Perilla trafo package. Please review manual and vignette documents of latest trafo package. The ATA.Transform function works with many different types of inputs.

ATA.Transform(X, tMethod = c("BoxCox", "BoxCox Shift", "Sqrt",
  "Sqrt Shift", "Reciprocal", "Log", "Log Shift", "NegLog", "Modulus",
  "Bickel-Doksum", "Manly", "Dual", "Yeo-Johnson", "GPower", "GLog"),
  tLambda, tShift = 0, bcMethod = c("loglik", "guerrero"),
  bcLower = 0, bcUpper = 1)

Arguments

X
a numeric vector or time series of class ts or msts for in-sample.

tMethod
Box-Cox power transformation family is consist of "BoxCox", "BoxCox Shift", "Sqrt", "Sqrt Shift", "Reciprocal", "Log", "Log Shift", "NegLog", "Modulus", "Bickel-Doksum", "Manly", "Dual", "Yeo-Johnson", "GPower", "GLog" in ATAforecasting package.

tLambda
Box-Cox power transformation family parameter. If NULL, data transformed before model is estimated.

tShift
Box-Cox power transformation family shifting parameter. If NULL, data transformed before model is estimated.

bcMethod
Choose method to be used in calculating lambda. "loglik" is default. Other method is "guerrero" (Guerrero, V.M. (1993)).

bcLower
Lower limit for possible lambda values. The lower value is limited by -5. Default value is 0.

bcUpper
Upper limit for possible lambda values. The upper value is limited by 5. Default value is 1.

Value

A list object consists of transformation parameters and transformed data. ATA.Transform is a list containing at least the following elements:

trfmX : Transformed data

tLambda : Box-Cox power transformation family parameter

tShift : Box-Cox power transformation family shifting parameter