Transformation Techniques for The ATAforecasting

The function provides the applicability of different types of transformation techniques for the data to which the Ata method will be applied. The ATA.Transform function works with many different types of inputs.

ATA.Transform(
  X,
  tMethod = c("Box_Cox", "Sqrt", "Reciprocal", "Log", "NegLog", "Modulus",
    "BickelDoksum", "Manly", "Dual", "YeoJohnson", "GPower", "GLog"),
  tLambda,
  tShift = 0,
  bcMethod = c("loglik", "guerrero"),
  bcLower = 0,
  bcUpper = 5
)

Arguments

X
a numeric vector or time series of class ts or msts for in-sample.

tMethod
Box-Cox power transformation family is consist of "Box_Cox", "Sqrt", "Reciprocal", "Log", "NegLog", "Modulus", "BickelDoksum", "Manly", "Dual", "YeoJohnson", "GPower", "GLog" in ATAforecasting package. If the transformation process needs shift parameter, ATA.Transform will calculate required shift parameter automatically.

tLambda
Box-Cox power transformation family parameter. If NULL, data transformed before model is estimated.

tShift
Box-Cox power transformation family shifting parameter. If NULL, data transformed before model is estimated.

bcMethod
Choose method to be used in calculating lambda. "loglik" is default. Other method is "guerrero" (Guerrero, V.M. (1993)).

bcLower
Lower limit for possible lambda values. The lower value is limited by -5. Default value is 0.

bcUpper
Upper limit for possible lambda values. The upper value is limited by 5. Default value is 1.

Value

A list object consists of transformation parameters and transformed data. ATA.Transform is a list containing at least the following elements:

trfmX : Transformed data

tLambda : Box-Cox power transformation family parameter

tShift : Box-Cox power transformation family shifting parameter

References

#'Tukey JW (1957). “On the Comparative Anatomy of Transformations.” The Annals of Mathematical Statistics, 28(3), 602--632.

#'Box GEP, Cox DR (1964). “An Analysis of Transformations.” Journal of the Royal Statistical Society. Series B (Methodological), 26(2), 211--252.

#'Manly BFJ (1976). “Exponential data transformations.” Journal of the Royal Statistical Society Series D, 25(1), 37--42.

#'John JA, Draper NR (1980). “An alternative family of transformations.” Journal of the Royal Statistical Society Series C, 29(2), 190--197.

#'Bickel PJ, Doksum KA (1982). “An analysis of transformations revisited.” Journal of the American Statistical Association, 76(374), 296--311.

#'Sakia RM (1992). “The Box-Cox Transformation Technique: A Review.” Journal of the Royal Statistical Society Series D, 41(2), 169--178.

#'Guerrero VM (1993). “Time-series analysis supported by power transformations.” Journal of Forecasting, 12(1), 37--48.

#'Yeo I, Johnson RA (2000). “A New Family of Power Transformations to Improve Normality or Symmetry.” Biometrika, 87(4), 954--959.

#'Durbin BP, Hardin JS, Hawkins DM, Rocke DM (2002). “A variance-stabilizing transformation for gene-expression microarray data.” Bioinformatics, 18(1), 105--110.

#'Whittaker J, Whitehead C, Somers M (2005). “The neglog transformation and quantile regression for the analysis of a large credit scoring database.” Journal of the Royal Statistical Society Series C, 54(4), 863--878.

#'Yang Z (2005). “A modified family of power transformations.” Economics Letters, 92(1), 14--19.

#'Kelmansky DM, Martinez EJ, Leiva V (2013). “A new variance stabilizing transformation for gene expression data analysis.” Statistical Applications in Genetics and Molecular Biology, 12(6), 653--666.